Add compatability note to README
Without this commit, the documentation in the README did not indicate
any compatability notes with a specific version of the API or Puppet
Enterprise. While the changes on some of the endpoints may not impact
this client, it is at least known that it will work with the API as
published in PE 2016.2.

This commit adds a simple compatibility note.