Use better names for settings
This commit uses a simpler method of passing settings to the client.
Originally it was built to accept port, API version, and hostname all
separately. It would then construct the proper URL internally. While
seemingly simple, it is not very straight forward. Additionally the
names used for the settings were cumbersome.

This commit uses a 'service-url' setting for passing the whole base URL
that will be constructed externally to this client. This commit also
updates all the documentation to use the new names for the settings.