Version set to 0.0.2
This commit will set the version of this gem to 0.0.2 so that it may be
published on rubygems.org. The original version slated for release,
0.0.1, had some lingering issues and was pulled shortly after
publishing. This version has applied the necessary changes, improvements,
and should server effectively as the first publicly available release.
Following this release, the gem will be strictly using semantic
versioning.