(maint) Move to OrchestratorClient
This renames the gem and module to orchestrator client. This gem is a
client to the orchestrator API not an implementation of the API.
Renaming it to orchestrator_client makes this clearer.

This removes access to the private "environment endpoints". These
endpoints are not considered public and so are removed from the Client
library

This standardizes config to make reading. This is easier to use if it
can read the client-tools config for orchestrator so some config values
are renamed.