Switch to using persistent HTTP connections
When running a series of very quick tasks with a short polling
interval the overhead of re-establishing the TCP connection
starts to show. On a series of test runs with 13 tasks less than
a second in duration with a 0.1s polling interval using a
persistent HTTP connection showed about a 8-9% performance
improvement.