(GH-2518) Add read-timeout configuration option
This adds a `read-timeout` configuration option which sets
`read_timeout` on the `Net::Http::Persistent` connection. This
configures the [seconds to wait until reading one
block](https://docs.seattlerb.org/net-http-persistent/Net/HTTP/Persistent.html#attribute-i-read_timeout),
which throws a `Net::ReadTimeout` error if there's nothing to read in
that time. This configuration option is exposed in Bolt when using the
PCP transport.

`Net::Http::Persistent#read_timeout=` should not be called if
`read-timeout` is not configured as there is no default for it (at least
that I could find).