(GH-29) Use regex to validate token file input
Previously we were using `URI.escape` to [verify that the provided
tokenfile path configuration didn't contain illegal
characters](https://github.com/puppetlabs/orchestrator_client-ruby/commit/3f100911622a94b80edebf8a976620b1edfc3993).
The method now raises a warning when called [as of
Ruby2.7](https://stackoverflow.com/questions/65423458/ruby-2-7-says-uri-escape-is-obsolete-what-replaces-it),
but the Orchestrator client was only using the function to verify that
no illegal characters were in the provided string. The way that
`URI.escape` worked [by default was using the
constant](https://stackoverflow.com/questions/34274838/why-is-uri-escape-marked-as-obsolete-and-where-is-this-regexpunsafe-constant)
`URI::UNSAFE` which is a regex to scan for illegal characters. Rather
than having to do [piecemeal URI
validation](https://docs.knapsackpro.com/2020/uri-escape-is-obsolete-percent-encoding-your-query-string)
we can simply use the `URI::UNSAFE` constant to check the string in the
same way we were before.

Closes #29